chore(spa-map): create station layer to reduce code duplication
